+++ R/clean.R
@@ -242,12 +242,7 @@ clean_recovery_msg <- function() {
   }
   # invoked manually in test-always-skipped.R
   # nocov start
-  message(
-    "Undo clean(garbage_collection = FALSE) with ",
-    "make(your_plan, recover = TRUE). ",
-    "Also builds unrecoverable targets. Message shown once per session ",
-    "if options(drake_clean_recovery_msg) is not FALSE."
-  )
+  cli_msg("Undo clean() with recover = TRUE in make().")
   .pkg_envir[["drake_clean_recovery_msg"]] <- FALSE
   invisible()
   # nocov end