+++ R/drake_config.R
@@ -306,7 +306,9 @@
 #'
 #' @param makefile_path Deprecated.
 #'
-#' @param console_log_file Optional character scalar of a file name or
+#' @param console_log_file Deprecated in favor of `log_make`.
+#'
+#' @param log_make Optional character scalar of a file name or
 #'   connection object (such as `stdout()`) to dump maximally verbose
 #'   log information for [make()] and other functions (all functions that
 #'   accept a `config` argument, plus `drake_config()`).
@@ -556,9 +558,10 @@ drake_config <- function(
   max_expand = NULL,
   log_build_times = TRUE,
   format = NULL,
-  lock_cache = TRUE
+  lock_cache = TRUE,
+  log_make = NULL
 ) {
-  logger <- logger(verbose = verbose, file = console_log_file)
+  logger <- logger(verbose = verbose, file = log_make)
   logger$disk("begin drake_config()")
   deprecate_fetch_cache(fetch_cache)
   deprecate_arg(hook, "hook") # 2018-10-25 # nolint
@@ -576,6 +579,7 @@ drake_config <- function(
   deprecate_arg(prepend, "prepend")
   deprecate_arg(makefile_path, "makefile_path")
   deprecate_arg(layout, "layout", "spec") # 2019-12-15
+  deprecate_arg(console_log_file, "console_log_file", "log_make") # 2020-02-08
   session_info <- resolve_session_info(session_info)
   memory_strategy <- match.arg(memory_strategy, choices = memory_strategies())
   if (memory_strategy == "memory") {